feat(api): pool-discriminator mining on from-intent discovery — frontend flows now produce pool questions (IND-418 follow-up)
The mining+question hook lived only on the MCP discovery-run queue, so
purely web-originated flows (intent creation/edit -> fromIntentQueue) never
mined a pool or asked a question - despite the intent page being P2's
surface. Extracted the hook into queues/pool/mining.shared.ts (trigger-
source-agnostic, same flags/budget/fire-and-forget contract) and wired it
into FromIntentQueue completion. DiscoveryRunQueue behavior unchanged (thin
delegate). Introducer flows still skipped; ad-hoc chat discovery without an
intentId remains shadow-only by design (pool questions are intent-scoped).

Also moves pool-question.shared.ts -> pool/question.shared.ts (no-dash
filename convention).

api 0.31.0. +2 from-intent tests (hook fires on completion; not on missing
intent).
