+++ services/api/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@indexnetwork/api",
-  "version": "0.30.0",
+  "version": "0.31.0",
   "license": "MIT",
   "private": true,
   "scripts": {
